Crate hungarian

Solve an assignment problem by Hungarian algorithm.

Example

Basic usage:

use hungarian::hungarian;

let result = hungarian(&[vec![2, 100, 10], vec![10, 100, 15]]);

assert_eq!(result.value, 17);
assert_eq!(&*result.forward, vec![0, 2].as_slice());
assert_eq!(&*result.backward, vec![Some(0), None, Some(1)].as_slice());

Structs

HungarianResult

A value object to represent the optimal solution of an assignment problem.

Traits

Value

A trait to adapt a value type to hungarian

Functions

hungarian

See the crate level documentation